backend/requirements.txt
# AgriIntel Demo — Python Dependencies
fastapi>=0.109
uvicorn[standard]>=0.27
numpy>=1.26
pandas>=2.1
rasterio>=1.3
pystac-client>=0.7
planetary-computer>=1.0
requests>=2.31
scikit-learn>=1.4
matplotlib>=3.8
xarray>=2024.1
openpyxl>=3.1          # for reading World Bank Excel sheets
Pillow>=10.2           # image manipulation for overlays
